requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
anthropic==0.43.0
supabase==2.10.0
python-dotenv==1.0.0
pydantic==2.5.3
httpx==0.26.0
openai
